Make BORG_BIN be the VENV bin.
Since the needed [virtualenv] and [virtualenvwrapper.sh] scripts are
already present in the BH_VENV environment that we create (and we
create it for the purposes of copying these files to BH_BIN), why
not just use $BH_ENV/bin as the BH_BIN? Just stict *that* directory
at the end of $PATH. It does not matter that we will have the sundry
other files from $BH_ENV/bin on the $PATH. We will install nothing
else to that environment. Any other activated venv will have access
to these scripts. And of another activated venv installs it's own
copy of these scripts, they will appear first in the $PATH.